+++ Sources/csuper-gui/main_menu.h
@@ -63,5 +63,6 @@ G_MODULE_EXPORT void changeDisplayDifferencePoints(GtkWidget *widget, gpointer d
 G_MODULE_EXPORT void printDialog(GtkWidget *widget, gpointer data);
 G_MODULE_EXPORT void beginPrint(GtkPrintOperation *operation, GtkPrintContext *context,gpointer user_data);
 G_MODULE_EXPORT void drawPrintPage(GtkPrintOperation *operation,GtkPrintContext *context,gint page_nr,gpointer user_data);
+G_MODULE_EXPORT void changeDisplayPointsGrid(GtkWidget *widget, gpointer data);
 
 #endif // MAIN_MENU_H_INCLUDED